Add all plugins needed by AdaptiveFeedbackCancellation.cfg
The goal for v4.5.2 is to have a working demo of adaptive feedback cancellation.
There is already an example in closed source MHA. This commit adds all the plugins
needed for the example configuration. I also copied over the config file and made
the changes needed to make it run standalone

Make openMHA compile on macOS

Adapt the configure script to work under macOS. Disable some clang-
specific warnings and fix a bug that prevented compilation.