Remove linkage of libopenmha fron unit-test-runners
On linux the unit-test-runner executables are not linked against libopenmha
despite the linker command requiring it. On macOS the linkage happens, this
causes errors as the library is not found at runtime. Why the behaviour of
clang and gcc differs is unclear for me. This solution is most likely a hack
and should be scrutinized as soon as possible.