Disable bad charset message in lint config
I removed the bad charset error because we have German umlauts
and the copyright symbol in every file with a copyright header.
This causes an unacceptable amount of false positives.

Setting the severity to "warning" or lower is not sufficient
because the lint engine completely stops after an instance of
"bad charset" is encountered.