Supress error message when receiving audio data while in state fw_stopping
Summary:
The iolib may continue to send data after it got a stop request
but before it processed it. Before, the framework would then complain
in its process call as process was called but it was not in running state.

As the state fw_stopping is active only in the time period after sending
a stop request to the io lib and it actually stopping, we may suppress
the error in that specific circumstance without harm.

Resolves T26.

Test Plan: Unit tests were successful; could no longer reproduce bug

Reviewers: tobias

Reviewed By: tobias

Maniphest Tasks: T26

Differential Revision: https://dev.openmha.org/D18