WIP: plugins: acPooling_wave: add initial probability bias
Add an initial implementation of a probability bias that can be
multiplied onto the pooled probability function in order to bias it
towards/against particular directions.  To this end a new configuration
variable "prob_bias" is added that represents the bias function.  The
biased probability function is then stored in the new AC variable
"prob_biased".

This is still considered WIP because it still needs some testing and
some aspects haven't been fully fleshed out yet (e.g., what the allowed
value range should be).