ac2lsl, ac2osc: Fix bug when checking for rt safety.
The previous implementation was buggy because static
variables in functions get shared along class instances as well
as functions calls. This is not thread-safe and introduces a bug
where rt-safety is only checked once when multiple instances of ac2x
are are used.

Replaced the static helper construct ensuring the rt check is only
run once by a call to std::run_once.