New test that exposes the bug that was closed by c69202a1
Summary:
MHA had a bug and disconnected some unrelated TCP clients
when one of the TCP clients closed its connection.

This bug was fixed in commit rOMDc69202a1.

This test fails at least sometimes when this bug is present in the MHA.

Test Plan:
* Execute the new test multiple times with the current MHA.
  * Verify the test succeeds every time.
* Execute the new test multiple times with an older version of openMHA (after commit 20af1419 but before commit c69202a1).
  * Verify that the test fails at least during some of the executions.

The test may not fail every time because the connections inside the MHA are
stored in a std::set of pointers, and only connections with higher memory
address value than the one that should be closed will also be closed
mistakenly. The memory addresses of the connection object inside the MHA are
unpredictable.

Maybe you can suggest a better test that will fail reliably.

Reviewers: pmaanen

Reviewed By: pmaanen

Subscribers: tobias

Differential Revision: https://dev.openmha.org/D32