Compiler identification
Summary:
Defines a macro COMPILER_ID that evaluates to a compile-time-constant string which identifies compiler, its version,
and the C++ standard, as "gcc-8.2.0-c++14" or "clang-6.0.0-c++11".

We want to include a variable that returns this string into every plugin, libopenmha.so, and mha.exe.
In future, MHA will refuse to load plugins compiled with different compilers. Plugins will refuse to link to
libopenmha compiled with different compiler.

No support for compilers other than gcc or clang, no support for C++ standards earlier than c++11.

Unit tests are provided for the building blocks.

On the way, I also fixed the compilation of external libraries: They are now compiled with the same compiler as the MHA.

Closes T307.

Test Plan:
* Apply Patch
* Compile MHA with gcc
* Execute unit-tests
  * Verify unit tests in directory mha/libmha succeed.
  * Watch out for additional printout in the form COMPILER_ID=... and verify that it is correct
* Repeat test with clang, preferably on Mac. I have done clang already on linux.

Reviewers: pmaanen

Reviewed By: pmaanen

Subscribers: tobias

Maniphest Tasks: T307

Differential Revision: https://dev.openmha.org/D36