Experimentally enable 2 cpu cores for compiling on ARM
ARM is the slowest platform to compile right now, even slightly slower than windows.
2 Build jobs for different ubuntu releases are running in parallel on a 4-core CPU.
This is attempt to assign 2 CPU cores to each build job to see if it makes a difference.
It could also be that the speed is IO-bound, reading and writing files to/from SD card.

Before this change, builds needed 50-53 minutes on ARM, vs 49-50 minutes on windows.