Add tooling for Windows installer creation
Summary:
To create the installer, in addition to the current prereqs nsis, dos2unix,
and the dependency walker have to be installed and present in the PATH.
I use nsis to create the installer. At the moment the user has to manually
run "make install" and then change into the ...packaging/exe/ directory and run make.

Test Plan: Works with MSYS2 on Win10 as described. Could install and start openMHA.

Reviewers: tobias, hendrikk

Reviewed By: tobias

Subscribers: jenkins-integration-bot, pmaanen

Maniphest Tasks: T271

Differential Revision: https://dev.openmha.org/D71