Fix response to carriage returns
Summary:
Previously, MHA responded with success to an empty line that ended
with a unix line-ending, and with failure to a line that ended with a
DOS line-ending.

This change aligns the behaviour so that both cases are successes.

It removes white space from end of line before parsing the command.

The change also adds a regression test.

Closes T436.

Test Plan:
Execute regression test on MHA before recompiling. Verify the test fails.

Compile MHA with the fix in place. Verify the test succeeds.

Reviewers: pmaanen

Reviewed By: pmaanen

Subscribers: jenkins-integration-bot, tobias, hendrikk

Maniphest Tasks: T436

Differential Revision: https://dev.openmha.org/D91