Correct level computation for spectral signals
Summary:
Correct level computation for fftfb_t library class and remove the correction
for the incorrect fftfb_t computation from multibandcompressor.

Add a test that checks the computed level in fftfbpow which was most severly
affected (3 dB error in level estimation) but not used anywhere in our examples.

Closes T573.

Test Plan:
Review. I have checked that the modified test fails with the incorrect level
computation but works with this correction as expected.

Reviewers: pmaanen, giso

Reviewed By: pmaanen

Maniphest Tasks: T573

Differential Revision: https://dev.openmha.org/D127