Change Jenkinsfiles to retire trusty, build PDFs, re-enable 32 bit deb files
Summary:
1) Trusty end-of-life was 2019-04-30.  Retire trusty builds.  Closes T640.

This change affects the Jenkinsfiles for building openMHA for branches
development and master, for the Review builds, and for precompiled external
libraries. A Makefile where Trusty was only used as an example in a comment
for an Ubuntu release name was also modified to not contain outdated release
names.

The documentation (Installation.md) is already free of Trusty references and
requires no changes.

2) Build PDF documentation as part of Jenkins automatic builds. PDFs are
built as part of the bionic 64 bit build and only published as Jenkins
artifacts for now. Ref T645.

3) 32 bits have been re-enabled because apt users see fewer warnings
when a repository for amd64 packages also distributes i386 packages,
and the new build server should allow the additional builds without
increased waiting time.

Test Plan:
I will kick off builds with these Jenkinsfiles with Jenkins' REPLAY option,
then link to these builds in comments on this Phabricator review request.

Check these builds and check the differences by review.

Reviewers: pmaanen, giso

Maniphest Tasks: T640, T645

Differential Revision: https://dev.openmha.org/D143