Shell skript to connect a Mahalia Beaglebone to the internet through USB
Summary:
Works when the BBB is attached to an Ubuntu PC. The BBB shares the
PC's internet connection and can install more packages via apt.
Created during and finished after Hackaton ref T683.

Test Plan:
- Test on a fresh Mahalia image on Ubuntu.
- Is this directory the correct directory to collect all these shell skripts?

I have tested this skript on ubuntu 18.04 and 19.04.
On 19.04, it works when the PC has bootet from kernel 4.15, but does not
work when the PC has bootet from kernel 5.0. The symptom is:
The USB network interface on the PC is recognized, the IP address 192.168.7.1
is assigned, but then the Beaglebone cannot be reached through this network
interface: ping 192.168.7.2 results in "cannot be reached". If you have any
idea, please share. Otherwise, this skript is useful enough on 18.04 IMO.

Reviewers: pmaanen, hendrikk

Maniphest Tasks: T683

Differential Revision: https://dev.openmha.org/D153