Call c++ parser in plugin without the detour through c
Summary:
This is a hackish solution for Ref T267. rOMDe96874f4196f sped up parsing by
45% on ARM. Profiling suggests that the strncpy in C interface parser is
responsible for 97% of the remaining CPU time spent in parsing.
This avoids the strncpy by using C++11 strings for data exchange
with plugins, relying on move semantics.

This also relies on all plugins compiled with exactly the same compiler
as MHA and libopenMHA, but does not check that this is the case. (It's a hack)

Closes T267.

Test Plan: Measure speed for parser-intensive MHA invocations before and after the change.

Reviewers: pmaanen, giso

Maniphest Tasks: T267

Differential Revision: https://dev.openmha.org/D161