Refer to getting-started-guide from README.md
Summary:
Closes T668. The original problem described by T668 was a combination of
these things:

- The README no longer told our users how to compile MHA and add it to the PATH, because
-- this has moved to the COMPILATION guide, and
-- we now have binary installers and an INSTALLATION guide.
- The example PATH had changes (Fixed during the 2nd coming of Ref T569)

The essence of this T668 that remained was that users needed to be
pointed to the some first steps, that we now provide with the getting-started
guide.

Therefore, instead of (incompletely) telling users how to execute some random
example in README.md, we now point to the getting-started-guide at some
URL where they can download it even before having installed openMHA.

I am using the gh-pages for storing the getting-started-guide at
http://www.openmha.org/docs/openMHA_starting_guide.pdf

The release script has been extended to remind the release agent to update the
getting-started guide at this location, and to not check in PDFs into
the source code git because we do not do this anymore.

Test Plan:
@hendrikk decides if it is ok to provide the getting-started-guide at this URL
via github-pages.
@pmaanen reviews the changes of the release skript

Reviewers: hendrikk, pmaanen

Reviewed By: hendrikk, pmaanen

Subscribers: hendrikk, pmaanen

Maniphest Tasks: T668, T569

Differential Revision: https://dev.openmha.org/D172