MHAIOFile: Fix unexpected sign flips
Summary:
When writing floating point values to integer files libsndfile expects
its input to be in the range [-1.0,1.0]. There was no place in mha where
this was enforced. This caused the strange sign flips we observer.
I activated the automatic clipping feature of libsndfile. This should
not be a big problem because we always write to the same audio format
we read from, i.e. if the output data is expected to be in the range
[-1.0,1.0] the input data should have been in the same range. To avoid
clipping the user has to use a floating point format. I added a warning
about this in the documentation. Resolves T703.

Test Plan: Run the new matlab test (test_mhaiofile.m). Test should fail w/o fix.

Reviewers: tobias

Reviewed By: tobias

Subscribers: hendrikk

Maniphest Tasks: T703

Differential Revision: https://dev.openmha.org/D173