Disclose installer crash on Windows 7
Summary:
Closes T736 as wontfix.

The openMHA windows installer crashes on some windows 7 machines
**after** a successful installation. openMHA is usable despite this
crash.

The crash is something that the NSIS installer causes after installing
everything that we asked it to instlal. Spending any more effort on
finding the actual cause in the NSIS source code or trying workarounds
by adding unrelated skripted stuff to the installer is too much wasted
effort considering the upcoming end-of-life of Windws 7 in January 2020,
and because the error is purely cosmetic.

Test Plan: Review.

Reviewers: hendrikk, pmaanen

Reviewed By: hendrikk

Maniphest Tasks: T736

Differential Revision: https://dev.openmha.org/D203