osc2ac: Modernize implementation. Closes T726.
Summary:
I adressed some of the grievances I had with the osc2ac implementation:
 - Remove commented out debug code as per out SOP
 - Use the libmha function mha_error instead of the handrolled debug macro
 - Avoid raw loops, use std::for_each instead
 - No raw pointers, use smart pointers instead

I think the changes, while relatively minor, still benefit the code. The downside to the last
change is that we now need c++14 for std::make_unique. I think we can live with that
as iirc we will stop supporting c++11 and older anyway.

As of now there is no automated test for osc2ac. Should we make it procedure that
plugins that get refactored need tests?

Test Plan: Manual code review only

Reviewers: tobias, giso

Reviewed By: tobias

Maniphest Tasks: T726

Differential Revision: https://dev.openmha.org/D265