Fix rohBeam compilation on Windows. Closes T863.
Summary:
-Revert "Revert "Add rohBeam plugin""
-rohBeam: Use std Bessel function instead of posix function.

"j0" is the Bessel function of the first kind. It is standardized
in the POSIX.1-2008 standard. Apparently mingw's implementation is
not completely conforming as it complains about it missing j0.
I therefore replaced the call to j0 with a call to a std::sph_bessel.
Downside: The Bessel fcts are c++17 and later only.

Test Plan: Compile on all supported os

Reviewers: tobias

Reviewed By: tobias

Maniphest Tasks: T863

Differential Revision: https://dev.openmha.org/D266