Ignore harmless exception caused by octave 4.11.0 + openjdk11 error
Summary:
Octave on ubuntu 20.04 raises an error on java runtime.exec /after/
successful execution of the child process.  We use this in `mha_start.m`.
This workaround checks if runtime.exec evokes this particular error and
ignores it.

Closes T962.

Test Plan: Have tested on a notebook running the (not-yet released) ubuntu 20.04.

Reviewers: pmaanen, hendrikk, MaxZ, zainsohail04

Reviewed By: zainsohail04

Subscribers: #openmha

Maniphest Tasks: T962

Differential Revision: https://dev.openmha.org/D315