Add lsl2ac plugin
Summary:
This plugin adds the first version of the lsl2ac plugin. It currently
only supports MHA_AC_FLOAT-type variables. Type conversions from other
types of stream should be handled in the background.

This is a very early version of the plugin and lacks many features,
notably real-time safety and recovery of streams. Only the latest
chunk of data will be pulled from lsl's internal buffers and stored
as ac variable.
If there is no new data, the ac variables get zero'ed.
If the stream is lost or times out, the ac variable zero'ed and will
not be updated any more.
This implementation should only be used for evaluation purposes and
is not intended for production use.

Test Plan: Code review, live test example 19

Reviewers: tobias, hendrikk

Reviewed By: tobias, hendrikk

Maniphest Tasks: T732

Differential Revision: https://dev.openmha.org/D293