Push branch development to github after each successful Jenkins build
Summary:
openMHA is an open source project.  We want that advanced users can
find the current status of the development branch on github, and not
restrict it to our internal git server.  This way, advanced users can
test bug fixes or new features as soon as they have been merged to
branch development.  Closes T804.

This change is a precondition for a planned change in the release
procedure, ref T951.  From release 4.12.0 on, we do not want to
hide individual changes in a squash commit to branch master.  Instead,
individual feature and bugfix merges to branch development will
be visible as steps also in branch master.

Test Plan:
Jenkins will *not* execute the changed Jenkinsfile automatically in
the Code Review builds.  Instead, we need to use Replay builds with
the altered Jenkinsfile.  In addition, the code review build will
*not* execute on branch development itself, therefore
1. The "when" condition needs to be altered for the code review builds
2. An additional "git checkout development" may be required.

The first successful code review build of this new feature will then
already push branch development to github, i.e. before the change is
reviewed and accepted.  This is fine because the content of what is
pushed to github is branch development with the old Jenkinsfile, not
the state still under code review.

We will need to add the ssh key of jenkinsmaster to the github
account that we will use for the push.

Reviewers: pmaanen, giso, MaxZ

Reviewed By: pmaanen

Subscribers: #openmha

Maniphest Tasks: T951, T804

Differential Revision: https://dev.openmha.org/D303