Fail reliably when a matlab test modifies git-controlled files
Summary:
The packaging process inserts "modified" into the name of the
generated debian packages if it finds that a git-controlled file
has changed before packaging starts.

Detect in Jenkins if this happens, and fail the Jenkins job,
but still make the build artifacts available for inspection.

Closes T1051.

Test Plan:
Test with a replay build in the code review builds.
Current git development, has this problem, because a matlab
test modifies a git-controlled sound file.

The detection here should raise an alarm

Reviewers: pmaanen, hendrikk, MaxZ, giso

Reviewed By: pmaanen

Maniphest Tasks: T1051

Differential Revision: https://dev.openmha.org/D368