This configuration file contains comments about how the adaptive feedback cancellation algorithm is configured for real-time processing. There is also a list of hardware components, on which the current configuration successfully runs.
Test Plan:
Read the comments relevant to how to adjust the delay
parameter for real-time processing (delay_w). If there is a
hardware setup available, on which you can test the configuration,
after having adjusted the delay, do so so as to verify that the
provided explanations are enough for the users to adjust the delay
themselves.

Reviewers: tobias, hendrikk, joma, kamil

Reviewed By: hendrikk

Subscribers: pmaanen

Maniphest Tasks: T600

Differential Revision: https://dev.openmha.org/D200