Add a unit test for example 09
Summary:
This commit adds a unit test for the localization example and a corresponding
MAT file with reference data.  The test itself reads in the example
configuration file and modifies it to use MHAIOFile instead of MHAIOJack.
After processing the audio file, the resulting MAT file is compared with a
reference.  With this test in place, the openMHA release protocol can be
further simplified.

This closes T942.

Test Plan:
run the test suite, maybe also double-check the reference MAT file
to be thorough, i.e., verify that it is plausible before we start relying on it
for future tests.  (I only verified that the test suite passes locally in
Octave 5.2.0.)

Reviewers: hendrikk, tobias

Reviewed By: hendrikk, tobias

Maniphest Tasks: T942

Differential Revision: https://dev.openmha.org/D349