Double-check tutorials on how to write plugins
Summary: We agreed that the only thing to be done for this task is to re-read section 2.2 and 2.2.1 of the plugin_developer_manual. In my eyes it is sufficient enough for new developers to create new plugins. Since there are also many examples it should not be too big a problem to dive into it.

Test Plan: nothing to be tested

Reviewers: pmaanen, tobias, hendrikk

Reviewed By: tobias

Maniphest Tasks: T331

Differential Revision: https://dev.openmha.org/D425