Add mhactl_java.jar file to javaclasspath when needed
Summary:
Instead of instructing users to add the jar file, add the
jar file to the dynamic class path when it is needed and
not yet in the class path.

Test Plan:
Search all documentation for mhactl_java.jar

Confirm that, with this change, manually adding the .jar
file with javaaddpath is no longer necessary.

Reviewers: pmaanen, tobias

Reviewed By: tobias

Subscribers: hendrikk

Maniphest Tasks: T1056

Differential Revision: https://dev.openmha.org/D378