Add new plugin complex_scale
Summary:
It is a test for new plugin complex_scale and to parse complex numbers with spaces.

In function parse_1_complex of the file libmha/src/mha_parser.cpp, usefulness of the code find_first_not_of("\t") was tested.
For this plugin, spaces in a complex number are handled well, so this code should stay in the original file.

Change-Id: Ifa0cf191f08a215fabcb6c3d07c90624d158c982

Test Plan: ./mha mhalib=complex_scale "mha.factor=( 1 + 5.6 i )" cmd=quit

Reviewers: tobias

Reviewed By: tobias

Differential Revision: https://dev.openmha.org/D427

Change-Id: Ie9dea1c75485ef139c3ca5ca7c28c9131f144619