plugins: Re-Throw exceptions thrown during prepare() after local cleanup
Some plugins did catch exceptions thrown during their prepare call
did their local cleanup as they should but did not rethrow the
exception. This causes the framework to be in a partially
invalid state when the start command is issues. I fixed this where
it was obiously a bug, but there are many more places that deserve a
closer look:

$~/.../mha/plugins:  grep -Inr "catch(" . | wc -l
      91

Closes T1357

Auditors: tobias