Fix openMHA matlab network connection management.
Summary:
Connection.set_address returns a boolean success flag.
This flag is false when the hostname could not be resolved
or the target port cannot be connected to.
Previously, the return value was not checked, this caused
followup errors later (null pointer error).
Closes T1336.

Test Plan:
Test steps to reproduce original problem on
https://forum.openmha.org/viewtopic.php?f=15&t=107&p=166&sid=a7735cf0f69e616d46f3263b8f7cfdd1#p166
With this fix in place, a human readable error message is produced instead
of the null pointer error.

Reviewers: pmaanen, tobias

Reviewed By: pmaanen

Subscribers: MaxZ

Maniphest Tasks: T1336

Differential Revision: https://dev.openmha.org/D492