Fix wrongfully accessing 2nd element of empty cell array
Summary:
In the file 'mhactl_java.m' the cell array variable `connections` is persistent over multiple function calls e.g. in our testing routine. After a certain timeout the variable will be reset to an empty state which (in our testing suite) might happen rather randomly depending on the performance of our servers. In a later function call the variable might be accessed again (to be precise: the element in the 2nd row) but the cell array is empty which leads to an error.
There are two lines where this might happen. I fixed this by checking if `connections` is empty before accessing it.
Another approach might be to:
- not empty connections but fill it with zeros or -1 etc. to show that the array contains no usefull information.
- set a longer timeout for `connections` to be retired but this might lead us to inaccurate guessing due to too many unpredictable factors.
- change the function call `retire_connections` and let it check whether `connections` is empty.
Closes T1447

Test Plan: Since it randomly happend during a testing routine and vanished without any changes I do not have a save testing plan and would look out if it appears again any time soon. But please comment on my other suggestions for fixing this error.

Reviewers: tobias, pmaanen

Reviewed By: tobias

Maniphest Tasks: T1447

Differential Revision: https://dev.openmha.org/D510