Print Info line in run_mha_tests.m
Summary: run_mha_tests.m does not tell you that it is running tests and since the tests take almost two minutes it may sometimes lead to the impression that something crashed. That is why I made it print a line telling the brave developer that something is going on.

Test Plan: execute run_mha_tests.m and see if the line is printed

Reviewers: pmaanen, hendrikk

Reviewed By: pmaanen

Maniphest Tasks: T1485

Differential Revision: https://dev.openmha.org/D551