Make generatemhaplugindoc work for io plugins
Summary:
The same problem as for analysemhaplugin applies. Because the IO plugins are not "proper" (in the sense of inherits from mha_plugin_t) plugins, the mhapluginloader used to load the plugin and parse its contents can not handle them. I added an abstraction layer between the latex generator and the plugin that, depending on the type of plugin either delegates to mhapluginloader for normal plugins or to fw_t for io plugins and provides the interface needed by the generator that is not provided by fw_t.
Right now the MHAPLUGIN_DOCUMENTATION macro does not work for io plugins, instead the help text is written into both the `short description` and `detailed description` fields. All io plugins are hard coded into the "io" category and are hard coded to be documented as accepting only wave to wave processing.

Test Plan:
* Verify output of genereratemhaplugindoc stays the same for non-io plugins
* Check out new section "io" in openMHA_plugins.pdf artefact

Reviewers: #openmha, tobias

Reviewed By: tobias

Subscribers: hendrikk, tobias

Maniphest Tasks: T1370

Differential Revision: https://dev.openmha.org/D497