Mark external_libs/include as system header directory
Summary:
ASIO triggers Wdeprecated-declarations in clang-11. I could fix this
issue by disabling the warning by hand but that would also disable
the warning for openMHA itself, this is also not a sustainable strategy
as we typically do not have the power to fix external_libs and disabling
warnings one by one when they pop up is tedious.

Therefore I marked the external_libs directory as system directory, which has some consequences
regarding the inclusion order, which in my estimation are not relevant for us,
and disables warnings for the headers in external_libs.

Closes T1598

Test Plan: * Observe succesful compilation

Reviewers: tobias

Reviewed By: tobias

Maniphest Tasks: T1598

Differential Revision: https://dev.openmha.org/D557