MHAIOPortAudio: Make output of Pa_GetStreamInfo() available in MHAIOPortAudio
Summary:
The contents of the PaStreamInfo struct are available as monitors in the stream_info
parser. I left out PaStreamInfo::structversion as it's of no information to the end user.
Due to the lack of a double precision monitor class I cast the values to float. As the latency
is given in seconds and ranges roughly from 0.001s to 1s this should not be a problem.

Closes T1588

Test Plan:
Code review
Matlab test

Reviewers: hendrikk, tobias

Reviewed By: tobias

Maniphest Tasks: T1588

Differential Revision: https://dev.openmha.org/D571