Remove obsolete chain_name parameter of plugin constructors
Summary:
The second (chain_name) parameter of the MHA plugin constructors
was not actually used before this change.  It always received the
same value, "(no chain)", and was only a historical remnant of early
MHA functionality. This change remove this parameter from all plugins,
from the plugin loader, plugin tests and documentation.

Before this change, openMHA plugins used varying C++ types for the
parameters of plugin constructors:

The first parameter was and is a C handle to the AC space for the
plugin (which is a struct). Before this change, plugins used
* `algo_comm_t`,
* `algo_comm_t &`, or
* `const algo_comm_t &`
as the C++ data type of the first constructor parameter.  This has
now been unified to always use `algo_comm_t`, a value type.  The last
parameter was always the configured name of the plugin: e.g. an
`mhachain` could load an `rmslevel` plugin either as
```
algos = [ ... rmslevel ... ]
```
or as
```
algos = [ ... rmslevel:MicLevel ...]
```
in the latter case, `MicLevel` was used as the name of the `rmslevel`
plugin in the MHA configuration tree and as a prefix for the AC
variables published by this `rmslevel` plugin instance.  Instead of
`MicLevel`, an arbitrary other word could have been used. The first case
is equivalent to specifying `rmslevel:rmslevel`.

In different plugins, the last parameter had data types

* `const char *`
* `std::string`
* `const std::string &`

In a unification effort, the data type of the last parameter has been changed to `const std::string &` in all plugins.

Also the parameter names have been unified: The first parameter's name is now `iac` in all plugins, the last parameter's name is now `configured_name` in all plugins. Also the spacing has been unified, it is now always:
```
constructor(algo_comm_t iac, const std::string & configured_name)
```

Closes T1637.

Test Plan:
* Compiles
* Unit tests successful
* System tests successful
* Step by step tutorial in developer manual contains the changes
* Code review

Reviewers: pmaanen

Reviewed By: pmaanen

Maniphest Tasks: T1637

Differential Revision: https://dev.openmha.org/D576