Enable detailed plugin documentation for io plugins
Summary:
I added code to io_lib_t that resolves the symbols defined by
the MHAPLUGIN_DOCUMENTATION macro and stores the contents in data members
if the symbols are present in the io library loaded. Also changed the
code in generatemhaplugindoc so that is uses the new detailed description if
its available and defaults to the old behavior if it's not so that we are not
forced to add detailed descriptions to all io plugins at once.

While this does not result in the cleanest code, it works for now and should not result in
problems as it only adds some symbols to the io libraries when its used.

I also added documentation to MHAIOFile. This documentation is used as test for the documentation macro and
should not be taken as finished. It's mainly copy pasted from the manual.

Closes T1593

Test Plan:
Code review
Review generated plugin documentation

Reviewers: tobias, MaxZ

Reviewed By: tobias, MaxZ

Maniphest Tasks: T1593

Differential Revision: https://dev.openmha.org/D582