Enhance robustness of test_mha_start_output_streams
Summary:
This test has failed twice in the past few days in the same line 35
on different systems.  Using sleep/pause to perform synchronization
is not acceptable in general, but it is in tests like this:  MHA
should print its output after startup in a reasonable time after all.
Closes T1675.

Test Plan:
Review only. We cannot force re-occurrence of the error with
reasonable effort.

Reviewers: pmaanen, MaxZ

Maniphest Tasks: T1675

Differential Revision: https://dev.openmha.org/D596