Create windows installer for nalnl2wrapper
Summary:
Create a separate windows installer for the nalnl2wrapper.
Users who want to use the NAL NL2 prescription rule with
closed-source MHA or openMHA can install this wrapper in
order to use this prescription rule, but they need to purchase
the NAL NL2 DLL from the NAL themselves.

The installer will always be generated by manual invocation.
Not integrated into the build system because for building,
the proprietary DLL is required.

Test Plan:
[ ] Test creation of installer and smoke test of installing
    it and invoking the nalnl2wrapper

Reviewers: pmaanen

Differential Revision: https://dev.openmha.org/D599