Create deb package for nalnl2wrapper depending on wine
Summary:
In order to create the deb package, the nalnl2wrapper.exe
command line wrapper needs to be compiled on windows and then
copied to directory mha/tools/fitting/NAL-NL2/packaging/deb
on a linux machine with mhamakedeb installed.  Shell script
create-nl2-wrapper-deb can be invoked in this directory to
create a debian package which needs to be manually included
into our apt repository.  The created deb is independent of
ubuntu release.  It depends on wine32:i386 and octave.
The wrapper will only work once the user has installed their
licensed copy of the NAL-NL2 library into the directory
/usr/share/nalnl2wrapper/.

Creating the deb package requires that the exe package is
available, the creation of which requires that the
proprietary NAL NL2 library is available.  Both cannot be
part of openMHA and therefore creating the deb is not
included in our build system but has to be performed manually.

Test Plan:
Have done these:
[X] Upload deb package to our apt repository
[X] Install it with apt
[X] Verify wine32 gets installed
[X] Store NAL-NL2 DLL where it is needed
[X] Check with mhagui_fitting_offline that fitting with NAL-NL2 works

Reviewers: pmaanen, MaxZ

Reviewed By: pmaanen

Differential Revision: https://dev.openmha.org/D600