Integrate shell script logic into Jenkinsfile
The detection of development in directory external_libs (and
sub-directories) in order to rebuild the external libraries when
necessary has moved from a shell script into the Jenkinsfile.
This allows using the ssh key for pulling and pushing the openMHA
repository because this job updates the external_libs_development
branch.