Fix openMHA Debian package dependency on liblsl
Summary:
As long as liblsl uses a different soname for each patch release,
openMHA needs to depend on the exact version of liblsl that was
installed during compilation.

To lessen the impact of any unsynchronized liblsl update in just
TASCAR or just openMHA, reduce the dependency on liblsl from
"Depends" to "Recommends".  We do not plan to perform any
unsynchronized updates in future, but this change to "Recommends"
would lessen the impact should such an update occur.

Remove outdated conflict against lsl-liblsl.

Closes T1731.

Test Plan:
Have executed `make deb` with this change, and the openMHA debian
package has correctly included `Recommends: liblsl (=  1.14.0-focal)`
as a result, which is the liblsl version currently installed where
I have performed the test.

Reviewers: giso

Reviewed By: giso

Maniphest Tasks: T1731

Differential Revision: https://dev.openmha.org/D623