Limit Jenkins build archive to keep only the last 7 days
Summary:
Options to discard old builds taken from
https://stackoverflow.com/a/44155346/9952529

Older builds in the same branch will be discarded when
a new build is executed. Even with this setting, Jenkins
will always keep the last successful build.

We need to limit the number of builds to keep in order to
save disk space on the Jenkins server, see T1741.

Test Plan:
Tested in a replay build of openMHA branch development:
| Before: | 10GB Disk usage | {F18516}
| After: | 2GB Disk usage | {F18517}

Reviewers: loth, MaxZ

Reviewed By: MaxZ

Differential Revision: https://dev.openmha.org/D627