Add new plugin wave2lsl
Summary:
This plugin generates a LSL stream from the incoming time domain
audio signal in the process callback and send this LSL stream as
a chunk without changing the number of channels and frame size.
The documentation of the wave2lsl plugin has been completed.

Closes T1762.

Test Plan:
A 4 channel audio input signal has been sent as an LSL stream and captured with the
LabRecorder software. The captured signals are identical with the sent signals.

Reviewers: tobias

Reviewed By: tobias

Maniphest Tasks: T1762

Differential Revision: https://dev.openmha.org/D640