Make ac2lsl respect the stride of AC variables.
Summary:
ac2lsl did not use the stride of an AC variable to determine
the correct number of channels for the outgoing stream, instead
it used the number of entries and generated one channel per entry.
While it got the correct samples into the network, this was
nevertheless wrong or a number of reasons:
- AC variables are allowed to change their size, ac2lsl could not handle that.
- save_wave, where the stride hat the clear meaning of "number of channels" did
  lose this important meta information.
This commit sets the number of channels equal to the stride of the AC var (if stride!=0)
and determines the correct number of samples to send right before sending.

Fixes T1766

Test Plan: Send different waves through save_wave and ac2lsl, observe number of channels

Reviewers: tobias, kamil

Reviewed By: tobias, kamil

Subscribers: #openmha

Maniphest Tasks: T1766

Differential Revision: https://dev.openmha.org/D647