Ensure that liblo and liblsl have been found as dependencies
Summary:
If liblo and liblsl have been found as dependencies to include in
the Windows installer, this must mean that plugins using liblo and
liblsl have been compiled and are also included in the installer.

This addition will prevent us from releasing a Windows installer
that lacks these plugins in future, because a build failure would
result from this change if these plugins were not included.

Closes T1768.  Closes T1769.

Test Plan: Review

Reviewers: pmaanen, MaxZ

Reviewed By: pmaanen

Maniphest Tasks: T1769, T1768

Differential Revision: https://dev.openmha.org/D662