Make time correction variable the same size as the timestamp variable
Summary:
Previously we inserted a time correction on every process() call, getting
a new time correction every five seconds. In the case where there are multiple timestamps
per process(), i.e. when there are more than one sample per channel, this makes retroactive
application of the time correction to the time stamps hard as it is not intuitive which time correction
should apply to which time stamp. This commit changes the implementation so that there is now
a time correction vector instead of one time correction. This wastes some space as there are
many rendundant entries in time correction, but the application of time correction is straight forward.

Closes T1797
Depends On D669.

Test Plan:
Manual review
Record with ac2xdf

Reviewers: tobias, steffen

Reviewed By: tobias

Maniphest Tasks: T1797

Differential Revision: https://dev.openmha.org/D683