Make nominal sampling rate configurable to adjust interpolation of time stamps in ac2lsl
Summary:
The previous calculation of the nominal sampling rate only worked
when there was one sample per channel per process(). This is
obviously not general. Since we made the change to respect
the number of channels in an AC variable, this lead to problems as
the timestamps are interpolated based on the nominal sampling rate.
Make the nominal sampling rate configurable by the user, using zero
(irregular) as default. Closes T1798

Depends On D683

Test Plan: Manual review

Reviewers: steffen, tobias

Reviewed By: tobias

Maniphest Tasks: T1798

Differential Revision: https://dev.openmha.org/D684